+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fixed missing search bar on mobile devices. remp/remp#932
 - Added support for Redis Sentinel cluster in the app configuration. remp/remp#1035
   - Added new `REDIS_SENTINEL_SERVICE` environment variable to configure name of the Sentinel service. If used, sentinel hosts are expected to be configured in comma-separated `REDIS_URL` environment variable. 
+- Added `article_external_id` into response of API call `/api/conversions`. remp/remp#1031
+- Added information about article's tags, authors and sections into API call `/api/articles`. remp/remp#1031
 
 ### [Mailer]